refactor code to be more organized.
