+++ volttron_installer/components/platform/__init__.py
@@ -1,0 +1,9 @@
+from .deployment_progress_dialog import deployment_progress_dialog
+from .pyenv_install_dialog import pyenv_install_dialog
+from .install_agent_dialog import install_agent_dialog
+from .remove_agent_dialog import remove_agent_dialog
+from .platform_tabs import platform_tabs
+from .logs_tab import logs_tab_content
+from .configuration_tab import configuration_tab_content
+from .status_tab import data_tab_content
+from .delete_platform_dialog import delete_platform_dialog
