+++ volttron_installer/state.py
@@ -12,6 +12,10 @@ from .utils.validate_content import check_json, check_csv, check_path, check_yam
 from .utils.create_csv_string import create_csv_string, create_and_validate_csv_string
 from .navigation.state import NavigationState
 from .backend.models import AgentType, HostEntry, PlatformConfig, PlatformDefinition, ConfigStoreEntry, AgentDefinition, CreatePlatformRequest, CreateOrUpdateHostEntryRequest, ToolRequest, BACnetDevice, BACnetWritePropertyRequest
+from .states.platform_deployment import PlatformDeploymentState
+from .states.platform_status import PlatformStatusState
+from .states.platform_agent import PlatformAgentState
+from .states.platform_logs import PlatformLogState
 from .utils.create_component_uid import generate_unique_uid
 from .utils.conversion_methods import csv_string_to_usable_dict
 from .utils.validate_content import check_json
@@ -332,20 +336,11 @@ async def __instances_from_api__() -> dict[str, Instance]:
 
     return instances
 
-class PlatformPageState(rx.State):
+class PlatformPageState(PlatformDeploymentState):
     #TODO once we save a platform, we create a routing id 
     # off of it's instance name, and redirect the user to 
     # platforms/x, maybe we might have to delete the old 
     # routing id
-    session_hydrated: bool = False
-    platforms: dict[str, Instance] = {
-        # "new": Instance(
-        #     host=HostEntryModelView(),
-        #     platform=PlatformModelView()
-        # )
-    }
-    _working_platform: Instance = Instance(host=HostEntryModelView(), platform=PlatformModelView())
-    list_of_agents: list[AgentModelView] = []
 
     _host_resolvable: bool = True
     _host_pinging: bool = False
@@ -356,43 +351,10 @@ class PlatformPageState(rx.State):
     # we can check if the host is reachable or not. if it is, we set this to true.
     _host_resolved: bool = False
 
-    # Deployment progress tracking
-    _is_deploying: bool = False
-    _deployment_status: str = "idle"  # idle, running, success, failed
-    _deployment_logs: list[str] = []
-    _current_task: str = ""
-    _deployment_progress: int = 0  # 0-100
-    _show_deployment_dialog: bool = False
-    _deployment_steps: list[dict] = []
-
-    # Pyenv install dialog
-    _show_pyenv_dialog: bool = False
-    _pyenv_installing: bool = False
-    _pyenv_error: str = ""
-
-    # Platform status tracking
-    _platform_status: dict = {}  # Stores PlatformDeploymentStatus data
-    _status_loading: bool = False
-    _last_refresh_request: float = 0.0  # Timestamp of last refresh request for debouncing
-    _status_error: str = ""
-    _last_status_check: str = ""
-    _starting_platform: bool = False  # True while starting VOLTTRON
-    _stopping_platform: bool = False  # True while stopping VOLTTRON
-    _status_check_in_progress: bool = False  # Lock to prevent concurrent status checks
+    # Deployment progress variables moved to PlatformDeploymentState
+    # Platform status track & Connection status tracking moved to PlatformStatusState
+    # Log viewing moved to PlatformLogState
     
-    # Connection status tracking
-    _connection_status: str = "unknown"  # connected, disconnected, checking, unknown
-    
-    # Log viewing
-    _platform_logs: str = ""
-    _logs_loading: bool = False
-    _log_font_size: int = 12  # Font size in pixels
-    _log_wrap: bool = True  # Whether to wrap long log lines
-    _tailing: bool = False  # Whether we're actively tailing logs
-    _connection_method: str = ""  # e.g., "SSH with key authentication"
-    _last_connection_check: str = ""
-    _connection_error: str = ""
-
     # Delete platform dialog
     _show_delete_dialog: bool = False
     _delete_remote_files: bool = False
@@ -410,34 +372,9 @@ class PlatformPageState(rx.State):
     _connect_ssh_user: str = ""
     _connect_ssh_port: str = "22"
 
-    # Install agent dialog
-    _show_install_agent_dialog: bool = False
-    _install_agent_mode: str = "catalog"  # "catalog" or "manual"
-    _install_agent_identity: str = ""
-    _install_agent_source: str = ""
-    _install_agent_start: bool = True
-    _installing_agent: bool = False
-    removing_agent_uuid: str = ""  # UUID of agent currently being removed
-    starting_agent_uuid: str = ""  # UUID of agent currently being started
-    stopping_agent_uuid: str = ""  # UUID of agent currently being stopped
-    _selected_catalog_agent: str = ""  # identity of selected agent from catalog
-    _show_remove_agent_dialog: bool = False
-    _agent_to_remove_uuid: str = ""
-    _agent_to_remove_name: str = ""
+    # Install agent dialog variables moved to PlatformAgentState
 
     # Vars
-    @rx.var(cache=True)
-    def current_uid(self) -> str:
-        return self.router.page.params.get("uid", "")
-
-    @rx.var
-    def working_platform(self) -> Instance:
-        self._working_platform = self.platforms.get(self.current_uid, Instance(host=HostEntryModelView(), platform=PlatformModelView()))
-        return self._working_platform
-
-    @rx.var(cache=True)
-    def in_file_platforms(self) -> list[Instance]:
-        return [instance for instance in self.platforms.values() if instance.platform.in_file]
     
     @rx.var
     def new_agents_list(self) -> list[str]:
@@ -448,35 +385,6 @@ class PlatformPageState(rx.State):
             return []
         # logger.debug(f"nah because new agents are : {[agent.identity for agent in working_platform.platform.agents.values() if not agent.is_new]}")
         return [agent.identity for agent in working_platform.platform.agents.values() if not agent.is_new]
-
-    @rx.var
-    def platform_title(self) -> str:
-        if self.current_uid == "":
-            return " "
-        working_platform: Instance | None = self.platforms.get(self.current_uid, None)
-        if working_platform is None:
-            return " "
-        else:
-            return working_platform.platform.safe_platform['config']['instance_name']
-
-    # === vars for platform details ===
-    @rx.var
-    def password_field(self) -> str:
-        if self.current_uid == "":
-            return ""
-        working_platform: Instance | None = self.platforms.get(self.current_uid, None)
-        if working_platform is None:
-            return ""
-        return working_platform.password
-
-    @rx.var
-    def platform_deployed(self) -> bool:
-        if self.current_uid == "":
-            return False
-        working_platform: Instance | None = self.platforms.get(self.current_uid, None)
-        if working_platform is None:
-            return False
-        return working_platform.deployed
 
     # === end of platform detail vars ===
 
@@ -1097,18 +1005,18 @@ class PlatformPageState(rx.State):
                 self._delete_remote_files = False
                 self._delete_confirmed = False
                 self._deleting_platform = False
-
-            # Navigate back to index and show success message
-            yield NavigationState.route_to_index()
-            if delete_remote:
-                yield rx.toast.success(f"Platform '{instance_name}' and remote files have been deleted")
-            else:
-                yield rx.toast.success(f"Platform '{instance_name}' has been deleted")
+                
+            yield rx.toast.success("Platform removed successfully")
+            yield rx.redirect("/instances")
+            
+            # Additional wait and redirect to ensure instances page refreshes
+            await asyncio.sleep(0.5)
+            yield rx.redirect("/instances")
 
         except Exception as e:
             async with self:
                 self._deleting_platform = False
-            yield rx.toast.error(f"Error deleting platform: {str(e)}")
+            yield rx.toast.error(f"Error during deletion: {str(e)}")
 
     @rx.event
     async def generate_new_platform(self):
@@ -1373,640 +1281,17 @@ class PlatformPageState(rx.State):
         if field == "volttron_version":
             logger.info(f"[CONFIG UPDATE] After update, volttron_version = '{working_platform.platform.config.volttron_version}'")
 
-    @rx.event(background=True)
-    async def handle_deploy(self):
-        working_platform: Instance = self.platforms[self.current_uid]
+    # Deployment methods moved to PlatformDeploymentState
 
-        # Show deployment dialog and initialize state
-        async with self:
-            self._show_deployment_dialog = True
-            self._is_deploying = True
-            self._deployment_status = "running"
-            self._deployment_logs = []
-            self._deployment_steps = []
-            self._current_task = "Saving platform configuration..."
-            self._deployment_progress = 0
-        yield
 
-        try:
-            # Save platform config to backend before deploying
-            # This ensures any UI changes (like volttron_version) are persisted
-            # Convert frontend model views to backend models
-            config_dict = working_platform.platform.config.to_dict()
-            logger.info(f"[DEPLOY SAVE] Config dict volttron_version: {config_dict.get('volttron_version', 'NOT FOUND')}")
-            logger.info(f"[DEPLOY SAVE] Full config_dict: {config_dict}")
+    # Status methods moved to PlatformStatusState
 
-            agents_dict = {}
-            for agent_id, agent_view in working_platform.platform.agents.items():
-                agent_data = agent_view.to_dict()
-                # AgentDefinition requires pypi_package or source - use source field
-                agents_dict[agent_id] = AgentDefinition(
-                    identity=agent_data["identity"],
-                    source=agent_data.get("source"),
-                    config=agent_data.get("config"),
-                    config_store={
-                        k: ConfigStoreEntry(**v) for k, v in agent_data.get("config_store", {}).items()
-                    },
-                    config_store_allowed=agent_data.get("config_store_allowed", True)
-                )
 
-            platform_request = CreatePlatformRequest(
-                host_id=working_platform.host.id,
-                config=PlatformConfig(**config_dict),
-                agents=agents_dict
-            )
-            logger.info(f"[DEPLOY SAVE] Saving platform with volttron_version: {platform_request.config.volttron_version}")
-            await update_platform(working_platform.platform.config.instance_name, platform_request)
-            logger.info(f"[DEPLOY SAVE] Platform saved successfully")
-            async with self:
-                self._current_task = "Deploying platform..."
-                self._deployment_progress = 5
-            yield
+    # Platform lifecycle methods moved to PlatformStatusState
 
-            deploy_task = asyncio.create_task(
-                deploy_platform(working_platform.platform.config.instance_name, working_platform.password)
-            )
 
-            # Poll deploy progress while deployment runs
-            while not deploy_task.done():
-                try:
-                    progress = await get_deploy_progress(working_platform.platform.config.instance_name)
-                    async with self:
-                        self._deployment_steps = progress.get("steps", [])
-                        self._deployment_progress = progress.get("progress", self._deployment_progress)
-                        self._current_task = progress.get("current_task", self._current_task)
-                        logs = progress.get("logs", [])
-                        if logs:
-                            self._deployment_logs = logs[-200:]
-                except Exception as poll_error:
-                    async with self:
-                        self._deployment_logs.append(f"Progress check failed: {poll_error}")
-                        self._deployment_logs = self._deployment_logs[-200:]
-                yield
-                await asyncio.sleep(1)
+    # Log methods moved to PlatformLogState
 
-            response = await deploy_task
-            response_data = response.json()
-
-            # Parse the output and extract task information
-            output = response_data.get("output", "")
-            tasks = response_data.get("tasks", [])
-
-            async with self:
-                for line in output.split('\n'):
-                    if line.strip():
-                        self._deployment_logs.append(line)
-                self._deployment_logs = self._deployment_logs[-200:]
-
-                # Mark as deployed
-                working_platform.deployed = True
-                logger.debug(f"response: {response_data}")
-
-                # Update deployment state to success
-                self._deployment_status = "success"
-                self._is_deploying = False
-                self._current_task = f"Deployment completed successfully ({len(tasks)} tasks executed)"
-                self._deployment_progress = 100
-            yield rx.toast.success("Deployed Successfully!")
-        except Exception as e:
-            logger.error(f"[DEPLOY] Error deploying platform {working_platform.platform.config.instance_name}: {e}")
-            error_text = str(e)
-            # Also check for detail attribute on ApiError - may be JSON
-            if hasattr(e, 'detail'):
-                raw_detail = e.detail
-                # Try to parse JSON response to get actual error message
-                try:
-                    parsed = json.loads(raw_detail)
-                    if isinstance(parsed, dict) and "detail" in parsed:
-                        error_text = parsed["detail"]
-                    else:
-                        error_text = raw_detail
-                except (json.JSONDecodeError, TypeError):
-                    error_text = raw_detail
-                logger.error(f"[DEPLOY] Error detail: {error_text}")
-
-            async with self:
-                # Update deployment state to failed
-                self._deployment_status = "failed"
-                self._is_deploying = False
-                self._current_task = f"Deployment failed"
-                self._deployment_logs.append(f"ERROR: {error_text}")
-                self._deployment_logs = self._deployment_logs[-200:]
-
-                # Check if this is a Python version issue - show pyenv prompt
-                if "Unsupported Python version" in error_text or "Python 3.10" in error_text:
-                    logger.info(f"[DEPLOY] Python version issue detected, showing pyenv prompt")
-                    self._show_pyenv_dialog = True
-                    self._pyenv_error = "Python 3.10 is required but not found on the remote host."
-            yield
-            yield rx.toast.error(f"There was an error deploying platform: {working_platform.platform.config.instance_name}")
-
-    @rx.event
-    def close_deployment_dialog(self):
-        """Close the deployment progress dialog"""
-        self._show_deployment_dialog = False
-
-    @rx.event
-    def close_pyenv_dialog(self):
-        self._show_pyenv_dialog = False
-        self._pyenv_error = ""
-
-    @rx.event(background=True)
-    async def handle_install_pyenv(self):
-        """Install Python 3.10 via pyenv on the remote host and create venv."""
-        working_platform: Instance = self.platforms[self.current_uid]
-        async with self:
-            self._pyenv_installing = True
-            self._pyenv_error = ""
-        yield
-
-        try:
-            result = await install_python310(working_platform.platform.config.instance_name)
-            logger.info(f"[PYENV] Install result: {result}")
-            
-            # Refresh platform data to get updated custom_python_path
-            await self.hydrate_state()
-            
-            async with self:
-                self._pyenv_installing = False
-                self._show_pyenv_dialog = False
-            yield rx.toast.success("Python 3.10 installed via pyenv. Click Deploy to continue.")
-        except Exception as e:
-            logger.error(f"[PYENV] Install failed: {e}")
-            async with self:
-                self._pyenv_installing = False
-                # Extract the detail from ApiError if available
-                error_msg = str(e)
-                if hasattr(e, 'detail'):
-                    error_msg = e.detail
-                self._pyenv_error = error_msg
-            yield rx.toast.error("Pyenv install failed. Check details in the dialog.")
-
-    @rx.event(background=True)
-    async def refresh_platform_status_debounced(self):
-        """Debounced version - waits 2 seconds before refreshing to batch multiple operations"""
-        import time
-        import asyncio
-
-        async with self:
-            self._last_refresh_request = time.time()
-            request_time = self._last_refresh_request
-
-        # Wait 2 seconds
-        await asyncio.sleep(2)
-
-        async with self:
-            # Only refresh if no newer request came in
-            if request_time == self._last_refresh_request:
-                yield PlatformPageState.refresh_platform_status()
-
-    @rx.event(background=True)
-    async def load_platform_status_background(self):
-        """Background version for initial page load - shows loading state immediately"""
-        # Set loading states immediately so UI shows spinners
-        async with self:
-            self._status_loading = True
-            self._connection_status = "checking"
-
-        # Now run the actual status checks (these will update state when done)
-        yield PlatformPageState.refresh_platform_status()
-        yield PlatformPageState.check_connection()
-
-    @rx.event
-    async def refresh_platform_status(self):
-        """Fetch the current status of the platform from the backend"""
-        if not self.current_uid or self.current_uid not in self.platforms:
-            return
-
-        # Check if another status check is already running - if so, skip this one
-        if self._status_check_in_progress:
-            logger.debug("Status check already in progress, skipping concurrent request")
-            return
-
-        # Set the lock
-        self._status_check_in_progress = True
-
-        working_platform: Instance = self.working_platform
-
-        # For new/unsaved platforms, don't try to fetch status
-        if working_platform.new_instance or not working_platform.platform.in_file:
-            self._platform_status = {
-                "platform_id": working_platform.platform.config.instance_name,
-                "state": "not deployed",
-                "host_configured": False,
-                "keys_verified": False,
-                "agents": {}
-            }
-            self._status_loading = False
-            self._status_check_in_progress = False
-            return
-
-        # Set loading state FIRST so UI shows "Checking..." immediately
-        self._status_loading = True
-        self._status_error = ""
-
-        try:
-            from datetime import datetime
-            status_response = await get_platform_status(working_platform.platform.config.instance_name)
-            self._platform_status = status_response.dict() if hasattr(status_response, 'dict') else status_response
-            self._last_status_check = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-            logger.debug(f"Platform status: {self._platform_status}")
-
-            # Auto-detect deployed status: if platform is running or deployed state detected,
-            # update the deployed flag (for platforms deployed before this fix)
-            detected_state = self._platform_status.get("state", "unknown")
-            if detected_state in ["running", "deployed"] and not working_platform.deployed:
-                logger.info(f"Auto-detecting deployed status for {working_platform.platform.config.instance_name}")
-                working_platform.deployed = True
-                # Persist the deployed flag to backend
-                try:
-                    await mark_platform_deployed(working_platform.platform.config.instance_name, True)
-                except Exception as mark_err:
-                    logger.warning(f"Could not persist deployed status: {mark_err}")
-        except Exception as e:
-            logger.error(f"Error fetching platform status: {e}")
-            self._status_error = str(e)
-            # If not marked as deployed and we get an error, show as not deployed
-            if not working_platform.deployed:
-                self._platform_status = {
-                    "platform_id": working_platform.platform.config.instance_name,
-                    "state": "not deployed",
-                    "host_configured": False,
-                    "keys_verified": False,
-                    "agents": {}
-                }
-            else:
-                self._platform_status = {
-                    "platform_id": working_platform.platform.config.instance_name,
-                    "state": "unknown",
-                    "host_configured": False,
-                    "keys_verified": False,
-                    "agents": {}
-                }
-        finally:
-            self._status_loading = False
-            self._status_check_in_progress = False  # Always release the lock
-
-    @rx.event
-    async def check_connection(self):
-        """Check the connection to the platform host"""
-        if not self.current_uid or self.current_uid not in self.platforms:
-            return
-
-        working_platform: Instance = self.working_platform
-
-        # Only check for saved platforms (not new/unsaved ones)
-        if working_platform.new_instance or not working_platform.platform.in_file:
-            self._connection_status = "unknown"
-            self._connection_method = "Platform not saved"
-            return
-        
-        # Skip check if platform isn't deployed or in error state
-        if not working_platform.deployed:
-            self._connection_status = "unknown"
-            self._connection_method = "Platform not deployed"
-            return
-
-        self._connection_status = "checking"
-
-        try:
-            from datetime import datetime
-            result = await check_platform_connection(working_platform.platform.config.instance_name)
-            
-            if result.get("connected", False):
-                self._connection_status = "connected"
-                self._connection_method = result.get("connection_method", "Unknown method")
-                self._connection_error = ""
-            else:
-                self._connection_status = "disconnected"
-                self._connection_error = result.get("error", "Connection failed")
-            
-            self._last_connection_check = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        except Exception as e:
-            logger.error(f"Error checking connection: {e}")
-            self._connection_status = "disconnected"
-            self._connection_error = str(e)
-
-    @rx.event(background=True)
-    async def handle_start_platform(self):
-        """Start the VOLTTRON platform"""
-        async with self:
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-
-            working_platform: Instance = self.working_platform
-
-            if not working_platform.deployed:
-                yield rx.toast.error("Platform must be deployed before starting")
-                return
-
-            # Show starting state
-            self._starting_platform = True
-
-        yield
-
-        try:
-            response = await start_platform(working_platform.platform.config.instance_name)
-            response_data = response.json() if hasattr(response, 'json') else response
-
-            async with self:
-                self._starting_platform = False
-                # Immediately update UI state (optimistic update)
-                self._platform_status["state"] = "running"
-
-            # Check if platform was already running
-            if response_data.get("already_running", False):
-                yield rx.toast.info("VOLTTRON is already running!")
-            else:
-                yield rx.toast.success("Platform started successfully!")
-
-            # Background refresh to confirm (don't block UI)
-            yield PlatformPageState.refresh_platform_status()
-
-        except ApiError as e:
-            async with self:
-                self._starting_platform = False
-
-            yield rx.toast.error(f"Failed to start platform: {e.detail}")
-        except Exception as e:
-            async with self:
-                self._starting_platform = False
-
-            yield rx.toast.error(f"Error starting platform: {str(e)}")
-
-    @rx.event(background=True)
-    async def handle_stop_platform(self):
-        """Stop the VOLTTRON platform"""
-        async with self:
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-
-            working_platform: Instance = self.working_platform
-
-            if not working_platform.deployed:
-                yield rx.toast.error("Platform must be deployed before stopping")
-                return
-
-            # Show stopping state
-            self._stopping_platform = True
-
-        yield
-
-        try:
-            await stop_platform(working_platform.platform.config.instance_name)
-
-            async with self:
-                self._stopping_platform = False
-                # Immediately update UI state (optimistic update)
-                self._platform_status["state"] = "deployed"
-
-            yield rx.toast.success("Platform stopped successfully!")
-
-            # Background refresh to confirm (don't block UI)
-            yield PlatformPageState.refresh_platform_status()
-
-        except ApiError as e:
-            async with self:
-                self._stopping_platform = False
-
-            yield rx.toast.error(f"Failed to stop platform: {e.detail}")
-        except Exception as e:
-            async with self:
-                self._stopping_platform = False
-
-            yield rx.toast.error(f"Error stopping platform: {str(e)}")
-
-    @rx.event(background=True)
-    async def fetch_platform_logs(self, lines: int = 100):
-        """Fetch VOLTTRON platform logs"""
-        async with self:
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-            
-            self._logs_loading = True
-            working_platform: Instance = self.working_platform
-            instance_name = working_platform.platform.config.instance_name
-            
-        yield
-        
-        try:
-            result = await get_platform_logs(instance_name, lines)
-            
-            async with self:
-                self._platform_logs = result.get("logs", "No logs available")
-                self._logs_loading = False
-                
-        except ApiError as e:
-            async with self:
-                self._platform_logs = f"Error fetching logs: {e.detail}"
-                self._logs_loading = False
-        except Exception as e:
-            async with self:
-                self._platform_logs = f"Error: {str(e)}"
-                self._logs_loading = False
-
-    @rx.event(background=True)
-    async def delete_platform_logs(self):
-        """Delete VOLTTRON platform logs"""
-        async with self:
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-            
-            self._logs_loading = True
-            working_platform: Instance = self.working_platform
-            instance_name = working_platform.platform.config.instance_name
-            
-        yield
-        
-        try:
-            result = await delete_platform_logs(instance_name)
-            
-            async with self:
-                self._platform_logs = "Log file deleted. Click 'Refresh Logs' to verify."
-                self._logs_loading = False
-                
-            yield rx.toast.success("Log file deleted successfully!")
-                
-        except ApiError as e:
-            async with self:
-                self._platform_logs = f"Error deleting logs: {e.detail}"
-                self._logs_loading = False
-            yield rx.toast.error(f"Failed to delete logs: {e.detail}")
-        except Exception as e:
-            async with self:
-                self._platform_logs = f"Error: {str(e)}"
-                self._logs_loading = False
-            yield rx.toast.error(f"Error deleting logs: {str(e)}")
-    
-    def increase_log_font_size(self):
-        """Increase log font size"""
-        self._log_font_size = min(32, self._log_font_size + 2)
-    
-    def decrease_log_font_size(self):
-        """Decrease log font size"""
-        self._log_font_size = max(8, self._log_font_size - 2)
-    
-    def toggle_log_wrap(self):
-        """Toggle log line wrapping"""
-        self._log_wrap = not self._log_wrap
-
-    @rx.var
-    def tailing(self) -> bool:
-        """Whether we're actively tailing logs"""
-        return self._tailing
-
-    @rx.event(background=True)
-    async def start_tailing(self):
-        """Start tailing logs with tail -f over SSH"""
-        global _tail_processes
-
-        async with self:
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-
-            if self._tailing:
-                return  # Already tailing
-
-            working_platform: Instance = self.working_platform
-            platform_id = working_platform.platform.config.instance_name
-            host = working_platform.host
-
-            # Get connection details
-            ansible_user = host.ansible_user
-            ansible_host = host.ansible_host
-            ansible_port = str(host.ansible_port)
-            volttron_home = host.volttron_home if host.volttron_home else "~/.volttron"
-
-            self._tailing = True
-            self._platform_logs = ""  # Clear existing logs
-
-        yield
-
-        try:
-            # Build SSH command for tail -f
-            ssh_cmd = [
-                "ssh",
-                "-o", "StrictHostKeyChecking=no",
-                "-o", "UserKnownHostsFile=/dev/null",
-                "-o", "BatchMode=yes",
-                "-p", ansible_port,
-                f"{ansible_user}@{ansible_host}",
-                f"tail -f {volttron_home}/volttron.log 2>/dev/null || echo 'Log file not found'"
-            ]
-
-            logger.debug(f"Starting tail process: {' '.join(ssh_cmd)}")
-
-            # Start the subprocess
-            process = subprocess.Popen(
-                ssh_cmd,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                bufsize=1  # Line buffered
-            )
-
-            _tail_processes[platform_id] = process
-
-            yield rx.toast.info("Started live log streaming...")
-
-            # Read lines in a loop
-            log_buffer = []
-            while True:
-                # Check if we should stop
-                async with self:
-                    should_continue = self._tailing
-
-                if not should_continue:
-                    break
-
-                # Check if process is still running
-                if process.poll() is not None:
-                    async with self:
-                        self._tailing = False
-                    yield rx.toast.warning("Log stream ended")
-                    break
-
-                # Try to read a line (non-blocking would be better but this works)
-                try:
-                    # Use select to check if data is available (Unix only)
-                    import select
-                    ready, _, _ = select.select([process.stdout], [], [], 0.5)
-
-                    if ready:
-                        line = process.stdout.readline()
-                        if line:
-                            log_buffer.append(line.rstrip('\n'))
-                            # Update state with new lines (batch updates for performance)
-                            if len(log_buffer) >= 1:
-                                async with self:
-                                    # Append new lines to existing logs
-                                    new_content = '\n'.join(log_buffer)
-                                    if self._platform_logs:
-                                        self._platform_logs = self._platform_logs + '\n' + new_content
-                                    else:
-                                        self._platform_logs = new_content
-                                    # Keep only last 1000 lines to prevent memory issues
-                                    lines = self._platform_logs.split('\n')
-                                    if len(lines) > 1000:
-                                        self._platform_logs = '\n'.join(lines[-1000:])
-                                log_buffer = []
-                                yield  # Update UI
-                    else:
-                        # No data available, just yield to allow UI updates
-                        await asyncio.sleep(0.1)
-
-                except Exception as e:
-                    logger.error(f"Error reading tail output: {e}")
-                    await asyncio.sleep(0.5)
-
-        except Exception as e:
-            logger.error(f"Error starting tail: {e}")
-            async with self:
-                self._tailing = False
-            yield rx.toast.error(f"Failed to start log streaming: {str(e)}")
-        finally:
-            # Clean up process
-            if platform_id in _tail_processes:
-                try:
-                    _tail_processes[platform_id].terminate()
-                    _tail_processes[platform_id].wait(timeout=2)
-                except:
-                    try:
-                        _tail_processes[platform_id].kill()
-                    except:
-                        pass
-                del _tail_processes[platform_id]
-
-    @rx.event
-    def stop_tailing(self):
-        """Stop tailing logs"""
-        global _tail_processes
-
-        self._tailing = False
-
-        # Get platform ID
-        if self.current_uid and self.current_uid in self.platforms:
-            working_platform: Instance = self.working_platform
-            platform_id = working_platform.platform.config.instance_name
-
-            # Kill the process if it exists
-            if platform_id in _tail_processes:
-                try:
-                    _tail_processes[platform_id].terminate()
-                    _tail_processes[platform_id].wait(timeout=2)
-                except:
-                    try:
-                        _tail_processes[platform_id].kill()
-                    except:
-                        pass
-                finally:
-                    if platform_id in _tail_processes:
-                        del _tail_processes[platform_id]
 
     @rx.event(background=True)
     async def handle_start_agent(self, agent_id: str):
@@ -2160,222 +1445,8 @@ class PlatformPageState(rx.State):
             yield rx.toast.error(f"Error stopping agent: {str(e)}")
 
     # Install agent dialog handlers
-    @rx.event
-    def open_install_agent_dialog(self):
-        """Open the install agent dialog"""
-        self._show_install_agent_dialog = True
-        self._install_agent_mode = "catalog"
-        self._install_agent_identity = ""
-        self._install_agent_source = ""
-        self._install_agent_start = True
-        self._selected_catalog_agent = ""
+    # Agent handling methods moved to PlatformAgentState
 
-    @rx.event
-    def close_install_agent_dialog(self):
-        """Close the install agent dialog"""
-        self._show_install_agent_dialog = False
-
-    @rx.event
-    def set_install_agent_mode(self, mode: str):
-        """Switch between catalog and manual mode"""
-        self._install_agent_mode = mode
-        # Clear selections when switching modes
-        self._selected_catalog_agent = ""
-        self._install_agent_identity = ""
-        self._install_agent_source = ""
-
-    @rx.event
-    def select_catalog_agent(self, agent_identity: str):
-        """Select an agent from the catalog"""
-        logger.debug(f"Selecting catalog agent: {agent_identity}")
-        self._selected_catalog_agent = agent_identity
-
-    @rx.event
-    def set_install_agent_identity(self, value: str):
-        self._install_agent_identity = value
-
-    @rx.event
-    def set_install_agent_source(self, value: str):
-        self._install_agent_source = value
-
-    @rx.event
-    def set_install_agent_start(self, value: bool):
-        self._install_agent_start = value
-
-    @rx.event(background=True)
-    async def handle_install_agent(self):
-        """Install an agent on the running platform"""
-        async with self:
-            # Get agent details based on mode
-            if self._install_agent_mode == "catalog":
-                # Find the selected agent from the catalog
-                selected_agent = None
-                for agent in self.list_of_agents:
-                    if agent.identity == self._selected_catalog_agent:
-                        selected_agent = agent
-                        break
-
-                if not selected_agent:
-                    yield rx.toast.error("Please select an agent from the catalog")
-                    return
-
-                agent_identity = selected_agent.identity
-                agent_source = selected_agent.source
-            else:
-                # Manual mode
-                if not self._install_agent_identity or not self._install_agent_source:
-                    yield rx.toast.error("Please enter agent identity and source")
-                    return
-
-                agent_identity = self._install_agent_identity
-                agent_source = self._install_agent_source
-
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-
-            working_platform: Instance = self.working_platform
-            self._installing_agent = True
-
-        yield
-
-        try:
-            await install_agent(
-                platform_id=working_platform.platform.config.instance_name,
-                agent_identity=agent_identity,
-                agent_source=agent_source,
-                start_agent=self._install_agent_start
-            )
-
-            async with self:
-                self._installing_agent = False
-                self._show_install_agent_dialog = False
-
-            yield rx.toast.success(f"Agent {agent_identity} installed successfully!")
-            yield PlatformPageState.refresh_platform_status()
-
-        except ApiError as e:
-            async with self:
-                self._installing_agent = False
-            yield rx.toast.error(f"Failed to install agent: {e.detail}")
-        except Exception as e:
-            async with self:
-                self._installing_agent = False
-            yield rx.toast.error(f"Error installing agent: {str(e)}")
-
-    @rx.event
-    def open_remove_agent_dialog(self, agent_uuid: str, agent_name: str):
-        """Open confirmation dialog before removing agent"""
-        logger.info(f"[REMOVE_DIALOG] Opening dialog for agent: {agent_name} ({agent_uuid})")
-        self._agent_to_remove_uuid = agent_uuid
-        self._agent_to_remove_name = agent_name
-        self._show_remove_agent_dialog = True
-        logger.info(f"[REMOVE_DIALOG] Dialog state set to: {self._show_remove_agent_dialog}")
-
-    @rx.event
-    def close_remove_agent_dialog(self, open_state: bool = False):
-        """Close the remove agent confirmation dialog"""
-        if not open_state:
-            self._show_remove_agent_dialog = False
-            self._agent_to_remove_uuid = ""
-            self._agent_to_remove_name = ""
-
-    @rx.event(background=True)
-    async def confirm_remove_agent(self):
-        """Actually remove the agent after confirmation"""
-        
-        # Check if status check is in progress to prevent VOLTTRON corruption
-        async with self:
-            if self._status_check_in_progress:
-                yield rx.toast.warning("System busy with status check, please wait...")
-                return
-            
-            self._status_check_in_progress = True
-            agent_uuid = self._agent_to_remove_uuid
-            
-            if not self.current_uid or self.current_uid not in self.platforms:
-                self._status_check_in_progress = False
-                yield rx.toast.error("Platform not found")
-                return
-
-            working_platform: Instance = self.working_platform
-            self.removing_agent_uuid = agent_uuid
-            self._show_remove_agent_dialog = False
-
-        yield
-
-        try:
-            await remove_agent(
-                platform_id=working_platform.platform.config.instance_name,
-                agent_uuid=agent_uuid
-            )
-
-            async with self:
-                self.removing_agent_uuid = ""
-                self._agent_to_remove_uuid = ""
-                self._agent_to_remove_name = ""
-                
-                # Optimistic update: remove from status dict if present
-                if self._platform_status and "agents" in self._platform_status:
-                    agents = self._platform_status["agents"]
-                    key_to_remove = None
-                    for k, v in agents.items():
-                        if v.get("uuid") == agent_uuid:
-                            key_to_remove = k
-                            break
-                    if key_to_remove:
-                        status = self._platform_status.copy()
-                        status["agents"] = {k: v for k, v in status["agents"].items() if k != key_to_remove}
-                        self._platform_status = status
-
-            yield rx.toast.success(f"Agent removed successfully!")
-            yield PlatformPageState.refresh_platform_status_debounced()
-
-        except ApiError as e:
-            async with self:
-                self.removing_agent_uuid = ""
-            yield rx.toast.error(f"Failed to remove agent: {e.detail}")
-        except Exception as e:
-            async with self:
-                self.removing_agent_uuid = ""
-            yield rx.toast.error(f"Error removing agent: {str(e)}")
-        finally:
-            async with self:
-                self._status_check_in_progress = False
-
-    @rx.event(background=True)
-    async def handle_remove_agent(self, agent_uuid: str):
-        """Remove/uninstall an agent from the running platform (deprecated - use open_remove_agent_dialog)"""
-        async with self:
-            if not self.current_uid or self.current_uid not in self.platforms:
-                yield rx.toast.error("Platform not found")
-                return
-
-            working_platform: Instance = self.working_platform
-            self._removing_agent = True
-
-        yield
-
-        try:
-            await remove_agent(
-                platform_id=working_platform.platform.config.instance_name,
-                agent_uuid=agent_uuid
-            )
-
-            async with self:
-                self._removing_agent = False
-
-            yield rx.toast.success(f"Agent removed successfully!")
-            yield PlatformPageState.refresh_platform_status_debounced()
-
-        except ApiError as e:
-            async with self:
-                self._removing_agent = False
-            yield rx.toast.error(f"Failed to remove agent: {e.detail}")
-        except Exception as e:
-            async with self:
-                self._removing_agent = False
-            yield rx.toast.error(f"Error removing agent: {str(e)}")
 
     @rx.event
     async def handle_save(self):
